fence_gce: Translate additional instance statuses
Map "STOPPING" to "on" since the instance is still powered on and may be
using shared resources.

Map "PROVISIONING" and "STAGING" to "off" since the instance has not
been booted yet and thus is not using shared resources.

Map "REPAIRING", "SUSPENDING", and "SUSPENDED" statuses to "unknown" and
log an error. These are not truly unknown statuses. A "REPAIRING" status
could result from an internal error or GCP maintenance issue. A
"SUSPEND(ING|ED)" status could result from user error. It seems improper
to only return an EC_STATUS error that tells the user, "Unable to obtain
correct plug status or plug is not available". However, we have to
return "on", "off", or "unknown", so I think the best approach would be
to return "unknown" and print an informative warning so that the user
knows how to correct it.

Reference:
  - https://cloud.google.com/compute/docs/instances/instance-life-cycle

Signed-off-by: Reid Wahl <nrwahl@protonmail.com>